* join Background and Delayed into one
