+++ src/MR.AspNetCore.Jobs/Server/InfiniteRetryProcessor.cs
@@ -26,7 +26,6 @@ namespace MR.AspNetCore.Jobs.Server
 				try
 				{
 					await _inner.ProcessAsync(context);
-					return;
 				}
 				catch (OperationCanceledException)
 				{
